feature: create subscriptionCosts component with styles and logic
