+++ src/features/profile/ui/settings/accountManagement/subscriptionCosts/index.ts
@@ -1,0 +1,1 @@
+export * from './subscriptionCosts'
