+++ src/features/profile/ui/settings/accountManagement/subscriptionCosts/subscriptionCosts.module.scss
@@ -1,0 +1,10 @@
+.costs {
+  order: 2;
+  width: 121.09%;
+}
+
+.container {
+  padding: 14px 20px;
+  background-color: var(--dark-500);
+  outline: 1px solid var(--dark-300);
+}
